refrac: Restructured library, for streamlined imports
