--- SSB_Spark_tools/utils/Usefull spark snippets.txt
@@ -1,2 +1,0 @@
-sc.getConf().getAll() # gets all spark session configs
-spark.read.path("/kilde*").show() # skows all data sets in a path
