--- SSB_Spark_tools/utils/proc_compare.py
@@ -1,2 +1,0 @@
-def foo(bar):
-    print(bar)
