+++ ssb_sparktools/__init__.py
@@ -1,5 +1,5 @@
 """SSB Spark Tools - Data proceeing tools developed and used by Statistics Norway (SSB)"""
 
-__version__ = '0.0.1'
+__version__ = '0.0.4'
 __author__ = 'Statistics Norway (ssb.no)'
 __all__ = []
