+++ ssb_sparktools/editing/__init__.py
@@ -1,0 +1,5 @@
+"""SSB Spark Tools - Data proceeing tools developed and used by Statistics Norway (SSB)"""
+
+__version__ = '0.0.4'
+__author__ = 'Statistics Norway (ssb.no)'
+__all__ = []
